+++ src/main/java/com/hexaware/policymanagement/repository/PolicyRepository.java
@@ -14,8 +14,8 @@ public interface PolicyRepository extends JpaRepository<Policy, Long>
 	
 	public List<Policy> findByCompany(String company);
 	
-	public List<Policy> findByPremiumGreaterThan(long premium);
+	public List<Policy> findByPremiumGreaterThan(double premium);
 	
-	public List<Policy> findByPremiumLessThan(long premium); 
+	public List<Policy> findByPremiumLessThan(double premium); 
 
 }
