+++ src/main/java/com/hexaware/policymanagement/services/IPolicyServices.java
@@ -12,8 +12,8 @@ public interface IPolicyServices
 	public void deleteByPolicyId(long policyId);
     public List<Policy> getPolicyByPolicyType(String policyType);
     public List<Policy> getPolicyByCompany(String company);
-    public List<Policy> getByPremiumLessThan(long premium);
-    public List<Policy> getByPremiumGreaterThan(long premium);
+    public List<Policy> getByPremiumLessThan(double premium);
+    public List<Policy> getByPremiumGreaterThan(double premium);
 
 	
 	public List<Policy> getAllPolicy();
